+++ backend/package.json
@@ -15,6 +15,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "mongoose": "^8.0.3",
     "nodemailer": "^6.9.7",
     "nodemon": "^3.0.2"
   }
