+++ backend/routes/contact.routes.js
@@ -1,0 +1,8 @@
+import express from "express";
+import sendEmail from "../controllers/sendEmail.js";
+
+const app = express.Router();
+
+app.post("/send-email", sendEmail);
+
+export default app;
